requirements.txt
asgiref==3.8.1
Django==4.2.11
pillow==10.3.0
python-barcode==0.15.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
